train
test data acc 0.88

+++ sleep_classification/data_loader.py
@@ -155,7 +155,7 @@ class EEGdataset(Dataset):
 
     def __getitem__(self,idx):
         x = self.data[idx]
-        y = self.label[idx]        
+        y = self.label[idx].long()        
         return x,y
